requirements.txt
pandas
streamlit
scikit-learn==1.0.2
streamlit_lottie==0.0.2
requests==2.24.0
Pillow==8.4.0
protobuf~=3.19.0
watchdog==2.1.8